requirements.txt
alembic==1.12.1
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
async==0.6.2
async-timeout==4.0.3
asyncpg==0.29.0
autoflake==2.2.1
backports.zoneinfo==0.2.1
bcrypt==4.1.1
billiard==4.2.0
black==23.12.1
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.104.1
fastapi-cache2==0.2.1
flake8==6.1.0
flower==2.0.1
greenlet==3.0.1
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
humanize==4.9.0
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
Mako==1.3.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
orjson==3.9.10
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
pendulum==2.1.2
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pyasn1==0.5.1
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic-extra-types==2.1.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyflakes==3.1.0
pyright==1.1.342
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
redis==4.6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqladmin==0.16.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
tornado==6.4
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
uvicorn==0.24.0.post1
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
WTForms==3.1.1
zipp==3.17.0
